Populate a queue of upcoming images to avoid repeating while loading
Loading the next image can often take a minute or more on particularly
slow machines. To work around this populate a queueu of images
(currently hardcoded to 6) which get populated and cycle through
regardless of what is happening with the refresh.

Additionally, we can use this cycling to kick off a refresh, instead
of having additionaly timers. Refresh is currently kicked off half way
through the queue and happens in the background (so cycling continues).